Revert "Added JWT Authentication."
