--- server/middleware/authMiddleWare.js
@@ -1,29 +1,0 @@
-const jwt = require('jsonwebtoken');
-require('dotenv').config();
-
-const Secret = process.env.SECRET;
-
-const verify = (req, res) => {
-
-    const token = req.cookies.jwt;
-
-    if(!token) {
-        res.status(401).json({ 
-            error: 'Unauthorized: No Token To Verify.' 
-        });
-    }
-    else {
-        jwt.verify(token, Secret, (err, validToken) => {
-            if(err) {
-                // console.log('invalid token', err.message);
-                res.status(401).json({ error: 'You are not an authenticated user' });
-            }
-            else {
-                // console.log('validToken - ', validToken);
-                res.status(200).json({ success: 'Successfully Authenticated. Happy Chating' })
-            }
-        });
-    }
-}
-
-module.exports = verify;
