+++ src/clientStuff/Message.js
@@ -3,7 +3,7 @@ import React from 'react'
 
 const Message = ({ data }) => { 
     return (
-        data.map(( { user, message, alignment, type }, index ) => {
+        data.map(( { user, message, alignment, type } , index ) => {
             return (
                 <div key={index} className={`chat-msg ${alignment}`} >
                     {type === 'message' ? <span>{user} : {message}</span> : <Emoji emoji={message} size={40} />}
