+++ src/clientStuff/helperFunctions.js
@@ -1,0 +1,10 @@
+import M from 'materialize-css'
+
+export const customAlert = (message) => {
+    M.toast({
+        html: message,
+        inDuration: 800,
+        outDuration: 800,
+        displayLength: 2000
+    })
+}
