--- src/components/emoji-picker/emojiStyle.css
@@ -1,33 +1,0 @@
-.reactions {
-    margin-top: 20px;
-}
-   
-.reactions .reactions-counter {
-    display: flex;
-    flex-wrap: wrap;  
-}
-   
-.reactions .emoji-mart {
-    z-index: 999;  
-}
-   
-.reactions .emoji-mart-anchors {
-    display: none;  
-}
-   
-.reactions .emoji-mart-search {
-    display: none;  
-}
-   
-.reactions .emoji-mart-bar {
-    display: none; 
-}
-   
-.reactions .emoji-mart-scroll {
-    overflow: scroll;
-    height: 8.5rem;  
-}
-   
-.reactions .emoji-mart-category-label {
-    display: none;
-}
