+++ src/components/majorComponents/ChatWindow.js
@@ -1,6 +1,6 @@
-import React, { useState, useEffect, useRef } from 'react'
-import { socket } from '../../client/Chat'
-import Message from '../../client/Message'
+import React,{ useState, useEffect, useRef } from 'react'
+import { socket } from '../../clientStuff/Chat'
+import Message from '../../clientStuff/Message'
 import chatTone from '../../../public/chat_tone.mp3'
 
 const ChatWindow = () => {
@@ -8,7 +8,7 @@ const ChatWindow = () => {
     const [ data, setData ] = useState([])
     const [ user, setUser ] = useState('')
     const [ CSS, setCSS ] = useState({
-        backgroundColor: '#c2c2c2', 
+        backgroundColor: '#c2c2c2',
         padding: '8px 20px'
     })
 
@@ -17,28 +17,25 @@ const ChatWindow = () => {
 
     useEffect(() => {
 
-        socket.on('myMsg', (newData) => {
+        socket.on('myMsg',(newData) => {
             setData(data => [ ...data, newData ])
         })
 
-        socket.on('chat', (newData) => {
+        socket.on('chat',(newData) => {
             setCSS({
-                ...CSS, 
-                display: 'none'
+                ...CSS,
+                display: 'none',
             })
             setUser('')
             setData(data => [ ...data, newData ])
-
-            if(audioRef){
-                audioRef.current.muted = false
-                audioRef.current?.play()
-            }
+            audioRef.current.muted = false
+            audioRef.current.play()
         })
 
-        socket.on('status', (user) => {
+        socket.on('status',(user) => {
             setCSS({
                 ...CSS,
-                display: 'block'
+                display: 'block',
             })
             setUser(user)
         })
