+++ webpack.config.js
@@ -9,8 +9,6 @@ module.exports = {
     },
     "devtool": "source-map",
     "devServer": {
-        historyApiFallback: true,
-        contentBase: './',
         port: 3000,
         open: true
     },
